leave makes user inactive (for non staff users. it keeps the relation to the program) fix rtl align for password_reset_email.html

+++ system/models.py
@@ -153,7 +153,10 @@ class Program(models.Model):
         return True
 
     def leave(self, user):
-        self.programuseraccess_set.filter(user=user).delete()
+        if user.is_staff or user.is_therapist:
+            self.programuseraccess_set.filter(user=user).delete()
+        else:
+            user.is_active = False
         return True
 
     @property
